# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SafarSathi carpooling API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rides:
    get:
      operationId: listRides
      tags: [rides]
      summary: List open rides feed
      parameters:
        - name: origin
          in: query
          schema:
            type: string
        - name: destination
          in: query
          schema:
            type: string
        - name: gender_preference
          in: query
          schema:
            type: string
            enum: [ANY, MALE, FEMALE]
        - name: date
          in: query
          description: Filter by departure date (YYYY-MM-DD)
          schema:
            type: string
        - name: origin_lat
          in: query
          description: Search origin latitude for proximity sort
          schema:
            type: number
        - name: origin_lng
          in: query
          description: Search origin longitude for proximity sort
          schema:
            type: number
        - name: dest_lat
          in: query
          description: Search destination latitude for proximity sort
          schema:
            type: number
        - name: dest_lng
          in: query
          description: Search destination longitude for proximity sort
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ride"

    post:
      operationId: createRide
      tags: [rides]
      summary: Post a new ride
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RideInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ride"
        "400":
          description: Bad request
...
  /rides/{id}:
    get:
      operationId: getRide
      tags: [rides]
      summary: Get ride details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ride"
        "404":
          description: Not found

    patch:
      operationId: updateRide
      tags: [rides]
      summary: Update a ride (driver only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RideUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ride"
        "404":
          description: Not found

    delete:
      operationId: deleteRide
      tags: [rides]
      summary: Cancel a ride (driver only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No content
  /rides/{id}/requests:
    get:
      operationId: listRideRequests
      tags: [requests]
      summary: List requests for a specific ride (driver only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RideRequest"
...
components:
  schemas:
...
    Ride:
      type: object
      required: [id, driver_id, driver_name, origin, destination, departure_time, available_seats, fare, transport_type, gender_preference, status, created_at]
      properties:
        id:
          type: integer
        driver_id:
          type: integer
        driver_name:
          type: string
        driver_university:
          type: string
        driver_avg_rating:
          type: ["number", "null"]
        origin:
          type: string
        destination:
          type: string
        origin_lat:
          type: ["number", "null"]
        origin_lng:
          type: ["number", "null"]
        dest_lat:
          type: ["number", "null"]
        dest_lng:
          type: ["number", "null"]
        departure_time:
          type: string
        available_seats:
          type: integer
        fare:
          type: integer
        transport_type:
          type: string
        gender_preference:
          type: string
          enum: [ANY, MALE, FEMALE]
        status:
          type: string
          enum: [OPEN, FULL, COMPLETED, CANCELLED]
        request_count:
          type: integer
        created_at:
          type: string
    RideInput:
      type: object
      required: [origin, destination, departure_time, available_seats, fare]
      properties:
        origin:
          type: string
          minLength: 1
        destination:
          type: string
          minLength: 1
        origin_lat:
          type: ["number", "null"]
        origin_lng:
          type: ["number", "null"]
        dest_lat:
          type: ["number", "null"]
        dest_lng:
          type: ["number", "null"]
        departure_time:
          type: string
        available_seats:
          type: integer
          minimum: 1
        fare:
          type: integer
          minimum: 0
        transport_type:
          type: string
        gender_preference:
          type: string
          enum: [ANY, MALE, FEMALE]
    RideUpdate:
      type: object
      properties:
        origin:
          type: string
        destination:
          type: string
        origin_lat:
          type: ["number", "null"]
        origin_lng:
          type: ["number", "null"]
        dest_lat:
          type: ["number", "null"]
        dest_lng:
          type: ["number", "null"]
        departure_time:
          type: string
        available_seats:
          type: integer
        fare:
          type: integer
        transport_type:
          type: string
        gender_preference:
          type: string
          enum: [ANY, MALE, FEMALE]
        status:
          type: string
          enum: [OPEN, FULL, COMPLETED, CANCELLED]
    RideRequest:
      type: object
      required: [id, ride_id, rider_id, rider_name, status, created_at]
      properties:
        id:
          type: integer
        ride_id:
          type: integer
        rider_id:
          type: integer
        rider_name:
          type: string
        rider_university:
          type: string
        rider_gender:
          type: string
        status:
          type: string
          enum: [PENDING, ACCEPTED, REJECTED]
        driver_phone:
          type: ["string", "null"]
          description: Revealed to the rider only when request is ACCEPTED
        rider_phone:
          type: ["string", "null"]
          description: Revealed to the driver only when request is ACCEPTED
        ride:
          $ref: "#/components/schemas/Ride"
        created_at:
          type: string
...