# lib/api-zod/src/generated/types/ride.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SafarSathi carpooling API
 * OpenAPI spec version: 0.1.0
 */
import type { RideGenderPreference } from './rideGenderPreference';
import type { RideStatus } from './rideStatus';

export interface Ride {
  id: number;
  driver_id: number;
  driver_name: string;
  driver_university?: string;
  /** @nullable */
  driver_avg_rating?: number | null;
  origin: string;
  destination: string;
  /** @nullable */
  origin_lat?: number | null;
  /** @nullable */
  origin_lng?: number | null;
  /** @nullable */
  dest_lat?: number | null;
  /** @nullable */
  dest_lng?: number | null;
  departure_time: string;
  available_seats: number;
  fare: number;
  transport_type: string;
  gender_preference: RideGenderPreference;
  status: RideStatus;
  request_count?: number;
  created_at: string;
}