# lib/api-zod/src/generated/types/rideInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SafarSathi carpooling API
 * OpenAPI spec version: 0.1.0
 */
import type { RideInputGenderPreference } from './rideInputGenderPreference';

export interface RideInput {
  /** @minLength 1 */
  origin: string;
  /** @minLength 1 */
  destination: string;
  /** @nullable */
  origin_lat?: number | null;
  /** @nullable */
  origin_lng?: number | null;
  /** @nullable */
  dest_lat?: number | null;
  /** @nullable */
  dest_lng?: number | null;
  departure_time: string;
  /** @minimum 1 */
  available_seats: number;
  /** @minimum 0 */
  fare: number;
  transport_type?: string;
  gender_preference?: RideInputGenderPreference;
}