# lib/api-zod/src/generated/types/rideUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SafarSathi carpooling API
 * OpenAPI spec version: 0.1.0
 */
import type { RideUpdateGenderPreference } from './rideUpdateGenderPreference';
import type { RideUpdateStatus } from './rideUpdateStatus';

export interface RideUpdate {
  origin?: string;
  destination?: string;
  /** @nullable */
  origin_lat?: number | null;
  /** @nullable */
  origin_lng?: number | null;
  /** @nullable */
  dest_lat?: number | null;
  /** @nullable */
  dest_lng?: number | null;
  departure_time?: string;
  available_seats?: number;
  fare?: number;
  transport_type?: string;
  gender_preference?: RideUpdateGenderPreference;
  status?: RideUpdateStatus;
}